financial_agent_crewai/web_app/backend/requirements.txt
langchain-chroma==0.2.2
langchain-core==0.3.41
langchain-huggingface==0.1.2
langchain-sambanova==0.1.2
langtrace-python-sdk==3.8.0
crewai==0.102.0
crewai-tools==0.36.0
Markdown==3.7
pandasai==2.4.2
pre-commit==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
sec-downloader==0.12.1
streamlit==1.42.2
weasyprint==64.1
yfinance==0.2.54

fastapi==0.115.8
uvicorn==0.34.0
starlette==0.45.3
pyjwt==2.10.1
cryptography==44.0.2
redis==5.2.1
